# backend/openapi_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: |
    api spec for StoryNet, covering the backend API and the internal apis of the pdf_service and ner_service. all public endpoints require JWT authentication unless explicitly marked otherwise.
  title: StoryNet API
  version: 2.0.0
servers:
#- description: backend API
#  url: https://angles-server.ddns.net:5001/api
- description: Internal pdf_service (OCR) — not publicly reachable
  url: http://pdf-service.app.svc.cluster.local:5001
- description: Internal ner_service (NER/LLM) — not publicly reachable
  url: http://ner-service.app.svc.cluster.local:5000
security:
- bearerAuth: []
...
paths:
...
  /documents/mine:
    get:
      description: |
        Returns only documents where the authenticated user appears in owner_ids.
      operationId: documents_mine_get
      parameters:
      - explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - queued
          - ocr_processing
          - ner_processing
          - completed
          - failed
          type: string
        style: form
      - explode: true
        in: query
        name: type
        required: false
        schema:
          enum:
          - text
          - pdf
          type: string
        style: form
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
          description: Paginated list of the user's own documents.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
      summary: List documents owned by the authenticated user
      tags:
      - Documents
      x-openapi-router-controller: openapi_server.controllers.documents_controller
  /documents/upload/pdf:
    post:
      description: |
        Accepts a PDF file. The backend forwards it to pdf_service for OCR, then pipes the page text to ner_service for entity extraction. Returns a job handle for polling.
      operationId: documents_upload_pdf_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/_documents_upload_pdf_post_request"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
          description: Job accepted.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unsupported media type (only application/pdf accepted).
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation error.
      summary: Upload a PDF document for OCR + NER processing
      tags:
      - Documents
      x-openapi-router-controller: openapi_server.controllers.documents_controller
  /documents/upload/text:
    post:
      description: |
        Accepts a plain-text file. The backend creates a Document record, enqueues a NER job, and returns the job handle. The caller can poll /jobs/{job_id} for status. The `public` field on the document is initialised from `make_public`.
      operationId: documents_upload_text_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/_documents_upload_text_post_request"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
          description: "Job accepted. Poll /jobs/{job_id} for progress."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation error.
      summary: Upload a plain-text document for NER processing
      tags:
      - Documents
      x-openapi-router-controller: openapi_server.controllers.documents_controller
  /documents/{id}:
    delete:
      description: |
        Removes the calling user from `owner_ids`. If no owners remain, the document and all orphaned entities are permanently deleted. If other owners remain, those owners and the document persist.
      operationId: documents_id_delete
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Ownership removed (or document deleted if last owner).
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authenticated but not permitted to access this resource.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Remove the authenticated user's ownership of a document
      tags:
      - Documents
      x-openapi-router-controller: openapi_server.controllers.documents_controller
    get:
      description: Returns 404 if the document is not visible to the authenticated
        user.
      operationId: documents_id_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Document metadata.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authenticated but not permitted to access this resource.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Get metadata for a specific document
      tags:
      - Documents
      x-openapi-router-controller: openapi_server.controllers.documents_controller
    patch:
      description: |
        Allows an owner to update mutable fields. The most important use case is releasing a document publicly (`public: true`). Any owner may set `public` to true; the change propagates to all entities and relationships whose sole source is this document. Only owners may call this endpoint.
      operationId: documents_id_patch
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_documents__id__patch_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Updated document metadata.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authenticated but not permitted to access this resource.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Update document metadata
      tags:
      - Documents
      x-openapi-router-controller: openapi_server.controllers.documents_controller
  /documents/{id}/entities:
    get:
      description: |
        Returns all FtM entities (node-type and interstitial) whose provenance includes this document. Visibility of the document itself must be satisfied first; individual entities further respect their own `public` flags.
      operationId: documents_id_entities_get
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Filter by FtM schema type.
        explode: true
        in: query
        name: schema
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntitySearchResponse"
          description: Paginated list of extracted entities.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid JWT.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Authenticated but not permitted to access this resource.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found or not visible to this user.
      summary: Get all entities extracted from a document
      tags:
      - Documents
      x-openapi-router-controller: openapi_server.controllers.documents_controller
...
components:
...
  schemas:
...
    Document:
      example:
        error_message: error_message
        entity_count: 5
        filename: filename
        public: true
        created_at: 2000-01-23T04:56:07.000+00:00
        id: id
        type: text
        owner_ids:
        - owner_ids
        - owner_ids
        status: queued
      properties:
        id:
          title: id
          type: string
        filename:
          title: filename
          type: string
        type:
          enum:
          - text
          - pdf
          title: type
          type: string
        status:
          enum:
          - queued
          - ocr_processing
          - ner_processing
          - completed
          - failed
          title: status
          type: string
        public:
          description: |
            True if one or more owners have released this document publicly. Any owner may set this to true. Once public it can only be reversed by all owners agreeing (business logic, not enforced here).
          title: public
          type: boolean
        owner_ids:
          description: IDs of all users who have uploaded this document.
          items:
            type: string
          title: owner_ids
          type: array
        entity_count:
          description: Number of FtM entities extracted from this document.
          readOnly: true
          title: entity_count
          type: integer
        created_at:
          format: date-time
          readOnly: true
          title: created_at
          type: string
        error_message:
          description: Present when status=failed.
          nullable: true
          title: error_message
          type: string
      required:
      - filename
      - id
      - owner_ids
      - public
      - status
      - type
      title: Document
      type: object
    JobStatus:
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        job_id: job_id
        progress: 8
        created_at: 2000-01-23T04:56:07.000+00:00
        document_id: document_id
        message: message
        status: queued
      properties:
        job_id:
          title: job_id
          type: string
        document_id:
          description: ID of the Document record created for this job.
          title: document_id
          type: string
        status:
          enum:
          - queued
          - ocr_processing
          - ner_processing
          - completed
          - failed
          title: status
          type: string
        progress:
          description: Estimated completion percentage.
          maximum: 100
          minimum: 0
          title: progress
          type: integer
        message:
          description: Human-readable status detail or error message.
          nullable: true
          title: message
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      required:
      - job_id
      - status
      title: JobStatus
      type: object
...
    EntitySearchResponse:
      example:
        total: 0
        page: 6
        results:
        - schema: Person
          provenance:
          - page_number: 1
            document_id: document_id
          - page_number: 1
            document_id: document_id
          first_seen: 2000-01-23T04:56:07.000+00:00
          public: false
          confidence: 0.87
          caption: caption
          id: id
          properties:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          owner_ids:
          - owner_ids
          - owner_ids
          last_changed: 2000-01-23T04:56:07.000+00:00
        - schema: Person
          provenance:
          - page_number: 1
            document_id: document_id
          - page_number: 1
            document_id: document_id
          first_seen: 2000-01-23T04:56:07.000+00:00
          public: false
          confidence: 0.87
          caption: caption
          id: id
          properties:
            name:
            - John Doe
            nationality:
            - us
            - au
            birthDate:
            - "1982"
          owner_ids:
          - owner_ids
          - owner_ids
          last_changed: 2000-01-23T04:56:07.000+00:00
        page_size: 1
      properties:
        total:
          title: total
          type: integer
        page:
          title: page
          type: integer
        page_size:
          title: page_size
          type: integer
        results:
          items:
            $ref: "#/components/schemas/FtMEntity"
          title: results
          type: array
      title: EntitySearchResponse
      type: object
    DocumentListResponse:
      example:
        total: 0
        page: 6
        results:
        - error_message: error_message
          entity_count: 5
          filename: filename
          public: true
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          type: text
          owner_ids:
          - owner_ids
          - owner_ids
          status: queued
        - error_message: error_message
          entity_count: 5
          filename: filename
          public: true
          created_at: 2000-01-23T04:56:07.000+00:00
          id: id
          type: text
          owner_ids:
          - owner_ids
          - owner_ids
          status: queued
        page_size: 1
      properties:
        total:
          title: total
          type: integer
        page:
          title: page
          type: integer
        page_size:
          title: page_size
          type: integer
        results:
          items:
            $ref: "#/components/schemas/Document"
          title: results
          type: array
      title: DocumentListResponse
      type: object
...
    ErrorResponse:
      example:
        details:
          key: ""
        error: ENTITY_NOT_FOUND
        message: message
      properties:
        error:
          description: Machine-readable error code.
          example: ENTITY_NOT_FOUND
          title: error
          type: string
        message:
          description: Human-readable description.
          title: message
          type: string
        details:
          additionalProperties: true
          nullable: true
          title: details
          type: object
      required:
      - error
      - message
      title: ErrorResponse
      type: object
...
    _documents_upload_text_post_request:
      properties:
        file:
          description: Plain-text file (UTF-8).
          format: binary
          type: string
        make_public:
          default: false
          description: |
            If true, the document is immediately marked public, causing all extracted entities and relationships to be publicly visible.
          type: boolean
      required:
      - file
      type: object
    _documents_upload_pdf_post_request:
      properties:
        file:
          description: PDF file.
          format: binary
          type: string
        make_public:
          default: false
          type: boolean
      required:
      - file
      type: object
    _documents__id__patch_request:
      properties:
        public:
          description: Set to true to release this document publicly.
          title: public
          type: boolean
      title: _documents__id__patch_request
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller.info_from_bearerAuth